feat: Add version checker and update notification
Implemented a version checker to compare the current version of vatACARS with the latest version available. If a newer version is available, an update notification dialog will be shown to the user. Also changed the base URL for HttpClientUtils to 'https://api.vatacars.com'. Added the necessary components and resources for the update notification dialog.