feat: add handling for Telex messages and Stations
Added handling for Telex messages and Station objects in the Tranceiver class. Also, added an interface for MessageData, which is implemented by CPDLCMessage and TelexMessage classes. This commit also includes changes in HoppiesInterface class to parse and handle CPDLC and Telex messages separately. Finally, a logger is added to log invalid CPDLCMessage format.