fix: update message list and modify button functionality
This commit includes a fix to update the message list in our vatACARS application. It also includes changes to the button functionalities in the EditorWindow.cs that help with the cloning process of UplinkEntries. Minor changes in positioning of the categories panel in the EditorWindow.Designer.cs have also been made.