feat: add CPDLC and Station messages for debugging
This commit introduces the addition of two CPDLC messages and a Station message in HoppiesInterface.cs for debugging purposes. The messages "TEST ONE" and "TEST TWO" have been added with a 'Y' response type. A new station with a callsign "VATACARS" has been added as well. Also, the method of adding messages to Tranceiver has been refactored.