feat: Added HandoffSelector component to vatACARS project
This commit introduces the HandoffSelector component to the vatACARS project. This involves changes to vatACARS.csproj, including the addition of HandoffSelector.cs and HandoffSelector.Designer.cs to the Compile Include. It also includes the addition of HandoffSelector.resx to the EmbeddedResource Include. the Util\LabelsXMLPatcher.cs and Util\Types.cs files have been included in the Compile Include.