fix: added event handler for station removal and improved message handling
Changes include addition of a new event handler for Station removal in Tranceiver class. Also, improved message handling by changing the way original messages are handled - added a new field 'Response' in the CPDLCMessage class to store response separately and updated message state only when the response is not 'STANDBY'. Also, fixed the delay time to be based on 'finishedMessageTimeout' setting.