refactor: changed datatype of selectedMsg to IMessageData
In this commit, the datatype of selectedMsg variable in PDCWindow.cs file has been changed from TelexMessage to IMessageData. This allows us to have more flexibility with the types of messages we can handle. Also, the method DateTime.Now has been replaced with DateTime.UtcNow for getting the time of message received, to ensure consistency across different time zones.