fix: Adjusted text display in EditorWindow component, resolves #106
This commit includes changes to the EditorWindow component in vatACARS. The 'this' keyword was removed when setting the Text property, and a new condition was added to an if statement to include messages with the state of 'ADSC'. Also, an Invalidate() call was added to the lbl_response control. The code changes aim to improve the clarity and accuracy.