feat: Add FDR variable to PDCWindow and QuickFillWindow
This commit includes the addition of a private FDR variable called networkPilotFDR in both the PDCWindow.cs and QuickFillWindow.cs files. This will provide a way for the system to track the flight data recorder information for the network pilot. Some unused imports in QuickFillWindow.cs were also removed to clean code.