Remove "STANDBY" message content assignment
Removed the line `message.Content = "STANDBY";` from the `stbyBtn.Click` event handler in `DispatchWindow.cs` within the `vatACARS.Components` namespace. This change ensures that the `Content` property of the `message` object is no longer set to "STANDBY" when the standby button is clicked. Other functionalities of the event handler remain intact.